requirements.txt
streamlit
flask
flask-cors
requests
python-dotenv
plotly
pandas